Show bio player only when video is actually ready
This should work around problems with the last video being shown on the current person by
only displaying the player at full opacity on the "loadeddata" event. This should be fine on
mobile, but we should double-check on iOS, to make sure that the event fires in time to show
the video in a user-triggerable state.